moved the collision logic for clarity
